requirements.txt
python=3.7.5
tqdm
sklearn
scikit-image
simpleitk
scipy
matplotlib

tensorflow-gpu==1.14.0
keras==2.2.4

torch==1.3.1
torchsummary